+++ tutorials/aio_echo/aio_echo_server.cpp
@@ -215,7 +215,7 @@ bool parseArguments(Params& _par, int argc, char* argv[])
 
 void Listener::onAccept(frame::aio::ReactorContext& _rctx, SocketDevice& _rsd)
 {
-    unsigned repeatcnt = backlog_size();
+    size_t repeatcnt = backlog_size();
 
     do {
         if (!_rctx.error()) {
